Add files via upload
MATLAB code for 1-D Richard equation using the h-based formulation, fully
implicit, tangent approximation.
Written by Saeed Kalantari and Danial Amini Baghbadorani
For the class of "Numerical Methods for Water Engineering", instructor:
Professor Ataie-Ashtiani, Spring 2018-2019
Reference paper:
Shahraiyni, H. T., & Ataie-Ashtiani, B. (2011). Mathematical forms and numerical schemes for the solution of unsaturated flow equations. Journal of Irrigation and Drainage Engineering, 138(1), 63-72.
https://ascelibrary.org/doi/abs/10.1061/%28ASCE%29IR.1943-4774.0000377

The Richards equation represents the movement of water in unsaturated soils and is attributed to Lorenzo A. Richards who published the equation in 1931. It is a nonlinear partial differential equation, which is often difficult to approximate since it does not have a closed-form analytical solution.